+++ RELEASE_NOTES.md
@@ -23,6 +23,7 @@ Release 3-dev
 - (wifi) Fix setting of stations' max supported channel width
 - (wifi) Fix setting TX power for HE TB PPDUs in case AP requested using the max TX power
 - (wifi) #521 - UL OFDMA support
+- (lr-wpan) #536 - Fixes PHY not going to TRX_OFF after CSMA-CA failure (RxOnWhenIdle(false))
 
 Release 3.36.1
 --------------